+++ sdk/python/tests/unit/infra/test_local_registry.py
@@ -1,4 +1,4 @@
-# Copyright 2021 The Feast Authors
+# Copyright 2022 The Feast Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
