+++ sdk/python/tests/unit/local_feast_tests/test_feature_service_apply.py
@@ -1,10 +1,7 @@
-import pytest
-
 from feast import FeatureService
 from tests.utils.cli_utils import CliRunner, get_example_repo
 
 
-@pytest.mark.integration
 def test_read_pre_applied() -> None:
     """
     Read feature values from the FeatureStore using a FeatureService.
